+++ testsuite/tests/codeGen/should_run/T2080.hs
@@ -10,7 +10,7 @@ utf8DecodeChar# :: Addr# -> Bool -> Bool
 {-# NOINLINE utf8DecodeChar# #-}
 utf8DecodeChar# a# fred =
   case () of
-    _ | isTrue# (word2Int# (indexWord8OffAddr# a# 0#) <=# 0x7F#) -> True
+    _ | isTrue# (word2Int# (word8ToWord# (indexWord8OffAddr# a# 0#)) <=# 0x7F#) -> True
 
 -- Omitting the next line gives an ASSERT error:
 -- ghc-6.9: panic! (the 'impossible' happened)