+++ testsuite/tests/ffi/should_run/PrimFFIWord8.hs
@@ -14,15 +14,15 @@ foreign import ccall "add_all_word8"
 
 main :: IO ()
 main = do
-    let a = narrowWord8# 0##
-        b = narrowWord8# 1##
-        c = narrowWord8# 2##
-        d = narrowWord8# 3##
-        e = narrowWord8# 4##
-        f = narrowWord8# 5##
-        g = narrowWord8# 6##
-        h = narrowWord8# 7##
-        i = narrowWord8# 8##
-        j = narrowWord8# 9##
-        x = W# (extendWord8# (add_all_word8 a b c d e f g h i j))
+    let a = wordToWord8# 0##
+        b = wordToWord8# 1##
+        c = wordToWord8# 2##
+        d = wordToWord8# 3##
+        e = wordToWord8# 4##
+        f = wordToWord8# 5##
+        g = wordToWord8# 6##
+        h = wordToWord8# 7##
+        i = wordToWord8# 8##
+        j = wordToWord8# 9##
+        x = W# (word8ToWord# (add_all_word8 a b c d e f g h i j))
     print x