+++ testsuite/tests/primops/should_run/T4442.hs
@@ -1,4 +1,6 @@
-{-# LANGUAGE MagicHash, UnboxedTuples, CPP #-}
+{-# LANGUAGE CPP #-}
+{-# LANGUAGE MagicHash #-}
+{-# LANGUAGE UnboxedTuples #-}
 module Main where
 
 #include "MachDeps.h"
@@ -8,32 +10,30 @@ import GHC.Stable(
   StablePtr(..), castStablePtrToPtr, castPtrToStablePtr, newStablePtr)
 import GHC.Exts
 import Data.Char(ord)
-#if WORD_SIZE_IN_BITS < 64
-import GHC.Int (Int64(..))
-import GHC.Word (Word64(..))
-#endif
+import GHC.Int
+import GHC.Word
 
-assertEqual :: (Show a, Eq a) => a -> a -> IO ()
-assertEqual a b
-  | a /= b = putStrLn (show a ++ " /= " ++ show b)
+assertEqual :: (Show a, Eq a) => String -> a -> a -> IO ()
+assertEqual msg a b
+  | a /= b = putStrLn (msg ++ " " ++ show a ++ " /= " ++ show b)
   | otherwise = return ()
 
-readBytes :: MutableByteArray# s -> State# s -> Int# -> (# State# s, [Int] #)
+readBytes :: MutableByteArray# s -> State# s -> Int# -> (# State# s, [Word8] #)
 readBytes marr s0 len = go s0 len []
  where
   go s 0# bs = (# s, bs #)
   go s i  bs = case readWord8Array# marr (i -# 1#) s of
-    (# s', b #) -> go s' (i -# 1#) (I# (word2Int# b):bs)
+    (# s', b #) -> go s' (i -# 1#) (W8# b : bs)
 
-indexBytes :: ByteArray# -> Int# -> [Int]
+indexBytes :: ByteArray# -> Int# -> [Word8]
 indexBytes arr len  =
-  [I# (word2Int# (indexWord8Array# arr i)) | I# i <- [0..I# len - 1]]
+  [W8# (indexWord8Array# arr i) | I# i <- [0..I# len - 1]]
 
-fillByteArray :: MutableByteArray# s -> Int# -> Int -> State# s -> State# s
-fillByteArray arr len (I# a) = go len
+fillByteArray :: MutableByteArray# s -> Int# -> Word8 -> State# s -> State# s
+fillByteArray arr len (W8# a) = go len
  where
   go 0# s = s
-  go i s = go (i -# 1#) (writeInt8Array# arr (i -# 1#) a s)
+  go i s = go (i -# 1#) (writeWord8Array# arr (i -# 1#) a s)
 
 test :: (Eq a, Show a)
   => String
@@ -43,7 +43,7 @@ test :: (Eq a, Show a)
   -> (MutableByteArray# RealWorld -> Int# -> a -> State# RealWorld
       -> State# RealWorld)
   -> a
-  -> [Int]
+  -> [Word8]
   -> Int
   -> IO ()
 test name index read write val valBytes len = do
@@ -53,14 +53,14 @@ test name index read write val valBytes len = do
   arrLen :: Int#
   arrLen = 24#
 
-  fillerByte :: Int
+  fillerByte :: Word8
   fillerByte = 0x34
 
-  expectedArrayBytes :: Int -> [Int]
+  expectedArrayBytes :: Int -> [Word8]
   expectedArrayBytes offset =
        replicate offset fillerByte
     ++ valBytes
-    ++ replicate (I# arrLen - len - offset) fillerByte
+    ++ replicate (fromIntegral $ I# arrLen - len - offset) fillerByte
 
   testAtOffset :: Int -> IO ()
   testAtOffset offset@(I# offset#) = runRW# (\s0 -> let
@@ -73,18 +73,18 @@ test name index read write val valBytes len = do
     actual1 = index arr offset#
     actualBytes1 = indexBytes arr arrLen
    in do
-       assertEqual actual0 val
-       assertEqual actual1 val
-       assertEqual actualBytes0 (expectedArrayBytes offset)
-       assertEqual actualBytes1 (expectedArrayBytes offset)
+       assertEqual "actual val 0" actual0 val
+       assertEqual "actual val 1" actual1 val
+       assertEqual "actualBytes0 indexed" actualBytes0 (expectedArrayBytes offset)
+       assertEqual "actualButes1 indexed" actualBytes1 (expectedArrayBytes offset)
    )
 
-intToBytes :: Int -> Int -> [Int]
-intToBytes (I# val0) (I# len0) = let
+intToBytes :: Word -> Int -> [Word8]
+intToBytes (W# val0) (I# len0) = let
     result = go val0 len0
     go v 0# = []
     go v len =
-      I# (v `andI#` 0xff#) : go (v `uncheckedIShiftRL#` 8#) (len -# 1#)
+      W8# (wordToWord8# v) : go (v `uncheckedShiftRL#` 8#) (len -# 1#)
   in
 #if defined(WORDS_BIGENDIAN)
     reverse result
@@ -92,28 +92,22 @@ intToBytes (I# val0) (I# len0) = let
     result
 #endif
 
-testIntArray ::
-     String
-  -> (ByteArray# -> Int# -> Int#)
+testIntArray :: (Eq a, Show a, Integral a, Num a)
+  => String
+  -> (ByteArray# -> Int# -> a)
   -> (MutableByteArray# RealWorld -> Int# -> State# RealWorld
-        -> (# State# RealWorld, Int# #))
-  -> (MutableByteArray# RealWorld -> Int# -> Int# -> State# RealWorld
-        -> State# RealWorld)
-  -> Int
+      -> (# State# RealWorld, a #))
+  -> (MutableByteArray# RealWorld -> Int# -> a -> State# RealWorld
+      -> State# RealWorld)
+  -> a
   -> Int
   -> IO ()
 testIntArray name0 index read write val0 len = do
   doOne (name0 ++ " positive") val0
   doOne (name0 ++ " negative") (negate val0)
  where
-  doOne name val = test
-    name
-    (\arr i -> I# (index arr i))
-    (\arr i s -> case read arr i s of (# s', a #) -> (# s', I# a #))
-    (\arr i (I# a) s -> write arr i a s)
-    val
-    (intToBytes val len)
-    len
+  doOne name val = test name index read write
+    val (intToBytes (fromIntegral val) len) len
 
 #if WORD_SIZE_IN_BITS == 64
 testInt64Array = testIntArray
@@ -143,24 +137,19 @@ testInt64Array name0 index read write val0 len = do
     len
 #endif
 
-testWordArray ::
-     String
-  -> (ByteArray# -> Int# -> Word#)
+testWordArray :: (Eq a, Show a, Integral a)
+  => String
+  -> (ByteArray# -> Int# -> a)
   -> (MutableByteArray# RealWorld -> Int# -> State# RealWorld
-        -> (# State# RealWorld, Word# #))
-  -> (MutableByteArray# RealWorld -> Int# -> Word# -> State# RealWorld
+        -> (# State# RealWorld, a #))
+  -> (MutableByteArray# RealWorld -> Int# -> a -> State# RealWorld
         -> State# RealWorld)
-  -> Word
+  -> a
   -> Int
   -> IO ()
-testWordArray name index read write val len = test
-  name
-  (\arr i -> W# (index arr i))
-  (\arr i s -> case read arr i s of (# s', a #) -> (# s', W# a #))
-  (\arr i (W# a) s -> write arr i a s)
-  val
-  (intToBytes (fromIntegral val) len)
-  len
+testWordArray name index read write val len =
+  test name index read write
+    val (intToBytes (fromIntegral val) len) len
 
 #if WORD_SIZE_IN_BITS == 64
 testWord64Array = testWordArray
@@ -203,7 +192,7 @@ float = 123.456789
 -- >>> import struct
 -- >>> import binascii
 -- >>> binascii.hexlify(struct.pack('>f', 123.456789))
-floatBytes :: Int
+floatBytes :: Word
 floatBytes = 0x42f6e9e0
 
 double :: Double
@@ -213,41 +202,61 @@ double = 123.45678901234
 -- >>> import struct
 -- >>> import binascii
 -- >>> binascii.hexlify(struct.pack('>d', 123.45678901234))
-doubleBytes :: Int
+doubleBytes :: Word
 doubleBytes = 0x405edd3c07fb4b09
 
 main :: IO ()
 main = do
   testIntArray "Int8#"
-    indexInt8Array# readInt8Array# writeInt8Array#
+    (\arr i -> I8# (indexInt8Array# arr i))
+    (\arr i s -> case readInt8Array# arr i s of (# s', a #) -> (# s', I8# a #))
+    (\arr i (I8# a) s -> writeInt8Array# arr i a s)
     123 1
   testIntArray "Int16#"
-    indexWord8ArrayAsInt16# readWord8ArrayAsInt16# writeWord8ArrayAsInt16#
+    (\arr i -> I16# (indexWord8ArrayAsInt16# arr i))
+    (\arr i s -> case readWord8ArrayAsInt16# arr i s of (# s', a #) -> (# s', I16# a #))
+    (\arr i (I16# a) s -> writeWord8ArrayAsInt16# arr i a s)
     12345 2
   testIntArray "Int32#"
-    indexWord8ArrayAsInt32# readWord8ArrayAsInt32# writeWord8ArrayAsInt32#
+    (\arr i -> I32# (indexWord8ArrayAsInt32# arr i))
+    (\arr i s -> case readWord8ArrayAsInt32# arr i s of (# s', a #) -> (# s', I32# a #))
+    (\arr i (I32# a) s -> writeWord8ArrayAsInt32# arr i a s)
     12345678 4
   testInt64Array "Int64#"
-    indexWord8ArrayAsInt64# readWord8ArrayAsInt64# writeWord8ArrayAsInt64#
+    (\arr i -> I64# (indexWord8ArrayAsInt64# arr i))
+    (\arr i s -> case readWord8ArrayAsInt64# arr i s of (# s', a #) -> (# s', I64# a #))
+    (\arr i (I64# a) s -> writeWord8ArrayAsInt64# arr i a s)
     1234567890123 8
   testIntArray "Int#"
-    indexWord8ArrayAsInt# readWord8ArrayAsInt# writeWord8ArrayAsInt#
+    (\arr i -> I# (indexWord8ArrayAsInt# arr i))
+    (\arr i s -> case readWord8ArrayAsInt# arr i s of (# s', a #) -> (# s', I# a #))
+    (\arr i (I# a) s -> writeWord8ArrayAsInt# arr i a s)
     int wordSizeInBytes
 
   testWordArray "Word8#"
-    indexWord8Array# readWord8Array# writeWord8Array#
+    (\arr i -> W8# (indexWord8Array# arr i))
+    (\arr i s -> case readWord8Array# arr i s of (# s', a #) -> (# s', W8# a #))
+    (\arr i (W8# a) s -> writeWord8Array# arr i a s)
     123 1
   testWordArray "Word16#"
-    indexWord8ArrayAsWord16# readWord8ArrayAsWord16# writeWord8ArrayAsWord16#
+    (\arr i -> W16# (indexWord8ArrayAsWord16# arr i))
+    (\arr i s -> case readWord8ArrayAsWord16# arr i s of (# s', a #) -> (# s', W16# a #))
+    (\arr i (W16# a) s -> writeWord8ArrayAsWord16# arr i a s)
     12345 2
   testWordArray "Word32#"
-    indexWord8ArrayAsWord32# readWord8ArrayAsWord32# writeWord8ArrayAsWord32#
+    (\arr i -> W32# (indexWord8ArrayAsWord32# arr i))
+    (\arr i s -> case readWord8ArrayAsWord32# arr i s of (# s', a #) -> (# s', W32# a #))
+    (\arr i (W32# a) s -> writeWord8ArrayAsWord32# arr i a s)
     12345678 4
   testWord64Array "Word64#"
-    indexWord8ArrayAsWord64# readWord8ArrayAsWord64# writeWord8ArrayAsWord64#
+    (\arr i -> W64# (indexWord8ArrayAsWord64# arr i))
+    (\arr i s -> case readWord8ArrayAsWord64# arr i s of (# s', a #) -> (# s', W64# a #))
+    (\arr i (W64# a) s -> writeWord8ArrayAsWord64# arr i a s)
     1234567890123 8
   testWordArray "Word#"
-    indexWord8ArrayAsWord# readWord8ArrayAsWord# writeWord8ArrayAsWord#
+    (\arr i -> W# (indexWord8ArrayAsWord# arr i))
+    (\arr i s -> case readWord8ArrayAsWord# arr i s of (# s', a #) -> (# s', W# a #))
+    (\arr i (W# a) s -> writeWord8ArrayAsWord# arr i a s)
     word wordSizeInBytes
 
   test
@@ -257,7 +266,7 @@ main = do
         case readWord8ArrayAsChar# arr i s of (# s', a #) -> (# s', C# a #))
     (\arr i (C# a) s -> writeWord8ArrayAsChar# arr i a s)
     'z'
-    [ord 'z']
+    [fromIntegral $ ord 'z']
     1
   test
     "WideChar#"
@@ -266,7 +275,7 @@ main = do
         case readWord8ArrayAsWideChar# arr i s of (# s', a #) -> (# s', C# a #))
     (\arr i (C# a) s -> writeWord8ArrayAsWideChar# arr i a s)
     '𠜎'  -- See http://www.i18nguy.com/unicode/supplementary-test.html
-    (intToBytes (ord '𠜎') 4)
+    (intToBytes (fromIntegral $ ord '𠜎') 4)
     4
   test
     "Addr#"
@@ -275,7 +284,7 @@ main = do
         case readWord8ArrayAsAddr# arr i s of (# s', a #) -> (# s', Ptr a #))
     (\arr i (Ptr a) s -> writeWord8ArrayAsAddr# arr i a s)
     (nullPtr `plusPtr` int)
-    (intToBytes int wordSizeInBytes)
+    (intToBytes word wordSizeInBytes)
     wordSizeInBytes
 
   stablePtr <- newStablePtr ()
@@ -288,7 +297,7 @@ main = do
     (\arr i p s -> case castPtrToStablePtr p of
                      (StablePtr a) -> writeWord8ArrayAsStablePtr# arr i a s)
     (castStablePtrToPtr stablePtr)
-    (intToBytes (castStablePtrToPtr stablePtr `minusPtr` nullPtr)
+    (intToBytes (fromIntegral $ castStablePtrToPtr stablePtr `minusPtr` nullPtr)
                 wordSizeInBytes)
     wordSizeInBytes